WIP: Rename `types.proto` to `mdc.proto`
