+++ test/screenshot/lib/cli.js
@@ -16,7 +16,7 @@
 
 'use strict';
 
-const mdcProto = require('../proto/types.pb').mdc.proto;
+const mdcProto = require('../proto/mdc.pb').mdc.proto;
 const {ApprovalId, DiffBase, GitRevision} = mdcProto;
 
 const argparse = require('argparse');
