+++ test/screenshot/lib/user-agent-store.js
@@ -16,7 +16,7 @@
 
 'use strict';
 
-const mdcProto = require('../proto/types.pb').mdc.proto;
+const mdcProto = require('../proto/mdc.pb').mdc.proto;
 const {UserAgent} = mdcProto;
 const {FormFactorType, OsVendorType, BrowserVendorType, BrowserVersionType} = UserAgent;
