+++ test/screenshot/proto/mdc.proto
